Refactoring structure of app packages.
