+++ app/src/main/java/com/gallopdevs/athanhelper/di/PrayerModule.kt
@@ -2,8 +2,8 @@ package com.gallopdevs.athanhelper.di
 
 import com.gallopdevs.athanhelper.model.PrayerCalculator
 import com.gallopdevs.athanhelper.model.PrayerCalculatorIpml
-import com.gallopdevs.athanhelper.model.PrayerRepo
-import com.gallopdevs.athanhelper.model.PrayerRepository
+import com.gallopdevs.athanhelper.repository.PrayerRepo
+import com.gallopdevs.athanhelper.repository.PrayerRepository
 import dagger.Binds
 import dagger.Module
 import dagger.hilt.InstallIn
