--- app/src/main/java/com/gallopdevs/athanhelper/home/MainActivityPagerAdapter.kt
@@ -1,18 +1,0 @@
-package com.gallopdevs.athanhelper.home
-
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.FragmentActivity
-import androidx.viewpager2.adapter.FragmentStateAdapter
-
-class MainActivityPagerAdapter(
-    fragmentActivity: FragmentActivity
-) : FragmentStateAdapter(fragmentActivity) {
-
-    private val fragmentList = ArrayList<Fragment>()
-
-    override fun createFragment(position: Int): Fragment = fragmentList[position]
-
-    override fun getItemCount(): Int = fragmentList.size
-
-    fun addFrag(fragment: Fragment) = fragmentList.add(fragment)
-}
