+++ app/src/main/java/com/gallopdevs/athanhelper/repository/PrayerRepository.kt
@@ -1,5 +1,6 @@
-package com.gallopdevs.athanhelper.model
+package com.gallopdevs.athanhelper.repository
 
+import com.gallopdevs.athanhelper.model.PrayerCalculator
 import javax.inject.Inject
 
 class PrayerRepository @Inject constructor(
