+++ app/src/main/java/com/gallopdevs/athanhelper/ui/clock/DayViewAdapter.kt
@@ -1,4 +1,4 @@
-package com.gallopdevs.athanhelper.clock
+package com.gallopdevs.athanhelper.ui.clock
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
