+++ app/src/main/java/com/gallopdevs/athanhelper/ui/settings/SettingsFragment.kt
@@ -1,4 +1,4 @@
-package com.gallopdevs.athanhelper.settings
+package com.gallopdevs.athanhelper.ui.settings
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,7 +7,6 @@ import android.view.ViewGroup
 import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.fragment.app.Fragment
-import com.gallopdevs.athanhelper.ui.settings.SettingsScreen
 import com.gallopdevs.athanhelper.ui.theme.AthanHelperTheme
 import dagger.hilt.android.AndroidEntryPoint
 import javax.inject.Inject
