+++ app/src/test/java/com/gallopdevs/athanhelper/viewmodel/ClockViewModelTest.kt
@@ -1,0 +1,6 @@
+package com.gallopdevs.athanhelper.viewmodel
+
+class ClockViewModelTest {
+
+
+}
